add: split components
